+++ src/components/Profile/EditProfile.tsx
@@ -1,207 +1,147 @@
-import { Fragment, useEffect, useState } from "react";
-import { Dialog, Transition } from "@headlessui/react";
-import { XMarkIcon } from "@heroicons/react/24/outline";
-import { useDispatch, useSelector } from "react-redux";
-import { RootState } from "../../app/store";
-
 const EditProfile = () => {
-  const [open, setOpen] = useState(true);
-  const dispatch = useDispatch();
-  const onEditProfilePage = useSelector(
-    (state: RootState) => state.switchOnOff
-  );
-  useEffect(() => {
-    setOpen(!open);
-  }, [onEditProfilePage]);
-
   return (
     <>
-      <Transition.Root show={open} as={Fragment}>
-        <Dialog as="div" className="relative z-10" onClose={setOpen}>
-          <Transition.Child
-            as={Fragment}
-            enter="ease-in-out duration-500"
-            enterFrom="opacity-0"
-            enterTo="opacity-100"
-            leave="ease-in-out duration-500"
-            leaveFrom="opacity-100"
-            leaveTo="opacity-0"
-          >
-            <div className="fixed inset-0 bg-gray-500 bg-opacity-75 transition-opacity" />
-          </Transition.Child>
+      <div
+        className="relative z-10"
+        aria-labelledby="modal-title"
+        role="dialog"
+        aria-modal="true"
+      >
+        <div className="fixed inset-0 bg-gray-500 bg-opacity-75 transition-opacity"></div>
 
-          <div className="pointer-events-none fixed inset-y-0 right-0 flex max-w-full pl-10">
-            <Transition.Child
-              as={Fragment}
-              enter="transform transition ease-in-out duration-500 sm:duration-700"
-              enterFrom="translate-x-full"
-              enterTo="translate-x-0"
-              leave="transform transition ease-in-out duration-500 sm:duration-700"
-              leaveFrom="translate-x-0"
-              leaveTo="translate-x-full"
-            >
-              <Dialog.Panel className="pointer-events-auto relative w-screen max-w-md">
-                <Transition.Child
-                  as={Fragment}
-                  enter="ease-in-out duration-500"
-                  enterFrom="opacity-0"
-                  enterTo="opacity-100"
-                  leave="ease-in-out duration-500"
-                  leaveFrom="opacity-100"
-                  leaveTo="opacity-0"
-                >
-                  <div className="absolute top-0 left-0 -ml-8 flex pt-4 pr-2 sm:-ml-10 sm:pr-4">
-                    <button
-                      type="button"
-                      className="rounded-md text-gray-300 hover:text-white focus:outline-none focus:ring-2 focus:ring-white"
-                      onClick={() => {
-                        setOpen(false);
-                      }}
-                    >
-                      <span className="sr-only">Close panel</span>
-                      <XMarkIcon className="h-6 w-6" aria-hidden="true" />
-                    </button>
-                  </div>
-                </Transition.Child>
-                <div className="flex h-full flex-col overflow-y-scroll bg-white py-6 shadow-xl">
-                  <div className="relative mt-6 flex-1 px-4 sm:px-6">
-                    {/* Replace with your content */}
-                    <div>
-                      <div className="mt-5 md:col-span-2 md:mt-0">
-                        <form action="#" method="POST">
-                          <div className="shadow sm:overflow-hidden sm:rounded-md">
-                            <div className="space-y-6 bg-white px-4 py-5 sm:p-6">
-                              <div className="grid grid-cols-3 gap-6">
-                                <div className="col-span-3 sm:col-span-2">
-                                  <label
-                                    htmlFor="company-website"
-                                    className="block text-sm font-medium text-gray-700"
-                                  >
-                                    Website
-                                  </label>
-                                  <div className="mt-1 flex rounded-md shadow-sm">
-                                    <span className="inline-flex items-center rounded-l-md border border-r-0 border-gray-300 bg-gray-50 px-3 text-sm text-gray-500">
-                                      http://
-                                    </span>
-                                    <input
-                                      type="text"
-                                      name="company-website"
-                                      id="company-website"
-                                      className="block w-full flex-1 rounded-none rounded-r-md border-gray-300 focus:border-indigo-500 focus:ring-indigo-500 sm:text-sm"
-                                      placeholder="www.example.com"
-                                    />
-                                  </div>
-                                </div>
-                              </div>
-
-                              <div>
-                                <label
-                                  htmlFor="about"
-                                  className="block text-sm font-medium text-gray-700"
-                                >
-                                  About
-                                </label>
-                                <div className="mt-1">
-                                  <textarea
-                                    id="about"
-                                    name="about"
-                                    rows={3}
-                                    className="mt-1 block w-full rounded-md border-gray-300 shadow-sm focus:border-indigo-500 focus:ring-indigo-500 sm:text-sm"
-                                    placeholder="you@example.com"
-                                    defaultValue={""}
-                                  />
-                                </div>
-                                <p className="mt-2 text-sm text-gray-500">
-                                  Brief description for your profile. URLs are
-                                  hyperlinked.
-                                </p>
-                              </div>
-
-                              <div>
-                                <label className="block text-sm font-medium text-gray-700">
-                                  Photo
-                                </label>
-                                <div className="mt-1 flex items-center">
-                                  <span className="inline-block h-12 w-12 overflow-hidden rounded-full bg-gray-100">
-                                    <svg
-                                      className="h-full w-full text-gray-300"
-                                      fill="currentColor"
-                                      viewBox="0 0 24 24"
-                                    >
-                                      <path d="M24 20.993V24H0v-2.996A14.977 14.977 0 0112.004 15c4.904 0 9.26 2.354 11.996 5.993zM16.002 8.999a4 4 0 11-8 0 4 4 0 018 0z" />
-                                    </svg>
-                                  </span>
-                                  <button
-                                    type="button"
-                                    className="ml-5 rounded-md border border-gray-300 bg-white py-2 px-3 text-sm font-medium leading-4 text-gray-700 shadow-sm hover:bg-gray-50 focus:outline-none focus:ring-2 focus:ring-indigo-500 focus:ring-offset-2"
-                                  >
-                                    Change
-                                  </button>
-                                </div>
-                              </div>
-
-                              <div>
-                                <label className="block text-sm font-medium text-gray-700">
-                                  Cover photo
-                                </label>
-                                <div className="mt-1 flex justify-center rounded-md border-2 border-dashed border-gray-300 px-6 pt-5 pb-6">
-                                  <div className="space-y-1 text-center">
-                                    <svg
-                                      className="mx-auto h-12 w-12 text-gray-400"
-                                      stroke="currentColor"
-                                      fill="none"
-                                      viewBox="0 0 48 48"
-                                      aria-hidden="true"
-                                    >
-                                      <path
-                                        d="M28 8H12a4 4 0 00-4 4v20m32-12v8m0 0v8a4 4 0 01-4 4H12a4 4 0 01-4-4v-4m32-4l-3.172-3.172a4 4 0 00-5.656 0L28 28M8 32l9.172-9.172a4 4 0 015.656 0L28 28m0 0l4 4m4-24h8m-4-4v8m-12 4h.02"
-                                        strokeWidth={2}
-                                        strokeLinecap="round"
-                                        strokeLinejoin="round"
-                                      />
-                                    </svg>
-                                    <div className="flex text-sm text-gray-600">
-                                      <label
-                                        htmlFor="file-upload"
-                                        className="relative cursor-pointer rounded-md bg-white font-medium text-indigo-600 focus-within:outline-none focus-within:ring-2 focus-within:ring-indigo-500 focus-within:ring-offset-2 hover:text-indigo-500"
-                                      >
-                                        <span>Upload a file</span>
-                                        <input
-                                          id="file-upload"
-                                          name="file-upload"
-                                          type="file"
-                                          className="sr-only"
-                                        />
-                                      </label>
-                                      <p className="pl-1">or drag and drop</p>
-                                    </div>
-                                    <p className="text-xs text-gray-500">
-                                      PNG, JPG, GIF up to 10MB
-                                    </p>
-                                  </div>
-                                </div>
-                              </div>
-                            </div>
-                            <div className="bg-gray-50 px-4 py-3 text-right sm:px-6">
-                              <button
-                                type="submit"
-                                className="inline-flex justify-center rounded-md border border-transparent bg-indigo-600 py-2 px-4 text-sm font-medium text-white shadow-sm hover:bg-indigo-700 focus:outline-none focus:ring-2 focus:ring-indigo-500 focus:ring-offset-2"
-                              >
-                                Save
-                              </button>
-                            </div>
+        <div className="fixed inset-0 z-10 overflow-y-auto">
+          <div className="flex min-h-full items-end justify-center p-4 text-center sm:items-center sm:p-0">
+            <div>
+              <div className="mt-5 md:col-span-2 md:mt-0">
+                <form action="#" method="POST">
+                  <div className="shadow sm:overflow-hidden sm:rounded-md">
+                    <div className="space-y-6 bg-white px-4 py-5 sm:p-6">
+                      <div className="grid grid-cols-3 gap-6">
+                        <div className="col-span-3 sm:col-span-2">
+                          <label
+                            htmlFor="company-website"
+                            className="block text-sm font-medium text-gray-700"
+                          >
+                            Website
+                          </label>
+                          <div className="mt-1 flex rounded-md shadow-sm">
+                            <span className="inline-flex items-center rounded-l-md border border-r-0 border-gray-300 bg-gray-50 px-3 text-sm text-gray-500">
+                              http://
+                            </span>
+                            <input
+                              type="text"
+                              name="company-website"
+                              id="company-website"
+                              className="block w-full flex-1 rounded-none rounded-r-md border-gray-300 focus:border-indigo-500 focus:ring-indigo-500 sm:text-sm"
+                              placeholder="www.example.com"
+                            />
                           </div>
-                        </form>
+                        </div>
+                      </div>
+
+                      <div>
+                        <label
+                          htmlFor="about"
+                          className="block text-sm font-medium text-gray-700"
+                        >
+                          About
+                        </label>
+                        <div className="mt-1">
+                          <textarea
+                            id="about"
+                            name="about"
+                            rows={3}
+                            className="mt-1 block w-full rounded-md border-gray-300 shadow-sm focus:border-indigo-500 focus:ring-indigo-500 sm:text-sm"
+                            placeholder="you@example.com"
+                            defaultValue={""}
+                          />
+                        </div>
+                        <p className="mt-2 text-sm text-gray-500">
+                          Brief description for your profile. URLs are
+                          hyperlinked.
+                        </p>
+                      </div>
+
+                      <div>
+                        <label className="block text-sm font-medium text-gray-700">
+                          Photo
+                        </label>
+                        <div className="mt-1 flex items-center">
+                          <span className="inline-block h-12 w-12 overflow-hidden rounded-full bg-gray-100">
+                            <svg
+                              className="h-full w-full text-gray-300"
+                              fill="currentColor"
+                              viewBox="0 0 24 24"
+                            >
+                              <path d="M24 20.993V24H0v-2.996A14.977 14.977 0 0112.004 15c4.904 0 9.26 2.354 11.996 5.993zM16.002 8.999a4 4 0 11-8 0 4 4 0 018 0z" />
+                            </svg>
+                          </span>
+                          <button
+                            type="button"
+                            className="ml-5 rounded-md border border-gray-300 bg-white py-2 px-3 text-sm font-medium leading-4 text-gray-700 shadow-sm hover:bg-gray-50 focus:outline-none focus:ring-2 focus:ring-indigo-500 focus:ring-offset-2"
+                          >
+                            Change
+                          </button>
+                        </div>
+                      </div>
+
+                      <div>
+                        <label className="block text-sm font-medium text-gray-700">
+                          Cover photo
+                        </label>
+                        <div className="mt-1 flex justify-center rounded-md border-2 border-dashed border-gray-300 px-6 pt-5 pb-6">
+                          <div className="space-y-1 text-center">
+                            <svg
+                              className="mx-auto h-12 w-12 text-gray-400"
+                              stroke="currentColor"
+                              fill="none"
+                              viewBox="0 0 48 48"
+                              aria-hidden="true"
+                            >
+                              <path
+                                d="M28 8H12a4 4 0 00-4 4v20m32-12v8m0 0v8a4 4 0 01-4 4H12a4 4 0 01-4-4v-4m32-4l-3.172-3.172a4 4 0 00-5.656 0L28 28M8 32l9.172-9.172a4 4 0 015.656 0L28 28m0 0l4 4m4-24h8m-4-4v8m-12 4h.02"
+                                strokeWidth={2}
+                                strokeLinecap="round"
+                                strokeLinejoin="round"
+                              />
+                            </svg>
+                            <div className="flex text-sm text-gray-600">
+                              <label
+                                htmlFor="file-upload"
+                                className="relative cursor-pointer rounded-md bg-white font-medium text-indigo-600 focus-within:outline-none focus-within:ring-2 focus-within:ring-indigo-500 focus-within:ring-offset-2 hover:text-indigo-500"
+                              >
+                                <span>Upload a file</span>
+                                <input
+                                  id="file-upload"
+                                  name="file-upload"
+                                  type="file"
+                                  className="sr-only"
+                                />
+                              </label>
+                              <p className="pl-1">or drag and drop</p>
+                            </div>
+                            <p className="text-xs text-gray-500">
+                              PNG, JPG, GIF up to 10MB
+                            </p>
+                          </div>
+                        </div>
                       </div>
                     </div>
+                    <div className="bg-gray-50 px-4 py-3 text-right sm:px-6">
+                      <button
+                        type="submit"
+                        className="inline-flex justify-center rounded-md border border-transparent bg-indigo-600 py-2 px-4 text-sm font-medium text-white shadow-sm hover:bg-indigo-700 focus:outline-none focus:ring-2 focus:ring-indigo-500 focus:ring-offset-2"
+                      >
+                        Save
+                      </button>
+                    </div>
                   </div>
-                  {/* /End replace */}
-                </div>
-              </Dialog.Panel>
-            </Transition.Child>
+                </form>
+              </div>
+            </div>
           </div>
-        </Dialog>
-      </Transition.Root>
+        </div>
+      </div>
     </>
   );
 };
